refactor: 랜덤 숫자 생성 함수를 util/random.js로 분리
- 기존 util 파일에 있던 getRandomNumber 함수를 util/random.js로 이동
- 관련 모듈에서 import 경로 업데이트
