+++ src/domain/RacingcarManager.js
@@ -1,5 +1,5 @@
 import Car from './Car.js';
-import { getRandomNumber } from '../util.js';
+import { getRandomNumber } from '../util/random.js';
 
 class RacingcarManager {
     createCars(carNames) {
